Bug fixes and improvements; rglButton added.
Fixed memory leaks in the rglGameStateMachine and implemented smart
pointers in the entire library to ensure deallocation.  Also invented a
state polling system in which a game state cannot change in the middle
of game state update.  Rather, checks for state changes are polled
immediately after the update process.  The reason I have done this is
because a game state could otherwise be exited during its own update
loop, therefore invalidating its contents, which are required to
complete the update.  Now, the user doesn't have to worry about a thing.
The game state machine will update when it feels is right.  Also added
an "unload" function in the texture manager so that textures can be
freed and save RAM.  Added an rglButton class for GUI implementation.