Much better FSM (final design... hopefully).
I think I finally got the whole FSM thing figured out.  With the new
implementation, the actual functions for state transitions don't touch
the vector of game states directly.  Rather, the transition is queued
into a separate vector, which includes each transition's type and game
state data.  Later, when the transitions are polled, the queued
transitions are simply iterated through and acted upon as they are
encountered.  This way, it is perfectly legal to push, change, or pop a
game state at any given point, even during state transitions.  One could
even call multiple state transitions in a row without polling and rest
assured that every transition will be completed later.  In addition,
this makes it nearly impossible (from what I've experienced) to crash
the FSM.  This will hopefully be the final design of the FSM, with
possible minor improvements in the future if needed.