Added functionality for state parsing and loading.
A system of parsing and loading state files instead of hard-coding them
has been implemented.  I created the rglObjectFactory class for object
creation based on a type attribute supplied in the state file.  I also
created the abstract rglObjectCreator class for defining how a certain
object should be created by the rglObjectFactory.  Added the pure
virtual function "load" to the rglGameObject class, which essentially
replaces the constructor.  This is used to provide a consistent method
of initialization for rglGameObjects.  Added the rglStateParser, which
essentially loads a state file and initializes the textures and objects
supplied.  Functionality for doing so comes from the TinyXML2 library.
Added an rglGUIState designed for states with buttons, etc.  Other minor
changes were made to other classes to support the new implementation.