Fixed a few things that needed attention.
Fixed an embarrassing bug in the TileLayer::update() function... Also
separated the GameStateMachine::update() function into two functions; I
moved the transition polling to pollTransitions(). The reasoning behind
this is that the clean() function needed to poll transitions without
necessarily updating the game. Weird behavior could have resulted
otherwise.

I also forgot to mention in the last commit that I changed the way
objects are loaded with the ObjectParams class. Instead of having a
bunch of member variables for every possible loader value an object
could have, all values are optional and stored in a map of strings.
Those strings can then be pulled by the loading object and automatically
converted to whatever data type. Pretty nifty, and cut my code down a
lot. Now we won't have to refractor practically every character when we
want to add a new parameter. It will be a whole lot easier.