Preparation for Box2D support.
Lots of changes were made to the way things work with this framework.
First, I added back the GameActor class and made GameObject a very
simple base class. My reasoning for this is that with Box2D, b2 objects
will work and behave much differently than conventional non-physics
objects. A simple base class provides this flexibility. The "load"
function that existed in GameObjects no longer exists; the constructor
has replaced it. This enables object creation on-the-fly without forcing
the user to instantiate a useless ObjectParams object. However, the
createObject function in the ObjectCreator base class essentially acts
as the new load function; it just interprets the values from the
ObjectParams object and passes them to the constructor. The difference
being ObjectParams is now only used for level loading, which is how it
was initially designed. Much more efficient than before. Each GameObject
also now has a pointer to the Level containing it. That way, the
GameObject can influence the Level and/or other GameObjects in the same
Level. The Level class now has an object creation/removal system similar
to that of the GameStateMachine. Most games involve some sort of dynamic
object creation and removal (bullets shooting, enemies dying/spawning,
etc.). In turn, I added onCreate and onDestroy virtual functions for the
GameObject class so the user can have control over what happens when the
object is actually added to the level (not necessarily when the object
is constructed). The LevelParser was redesigned and doesn't require
private access to the Level class. The framework is just about ready to
have physics implemented :D