Changed the way debugging works.
Okay... so here's the story. Box2D legit doesn't like debug
configuration. And there seems to be no way around that. So what I'm
doing is essentially requiring the user to use release configuration in
their game. But don't freak out! You can still kind of "fake" debugging.
There's a new parameter in Game::run that allows you to specify if you
want to use debugging. If that parameter is true, a console window pops
up and stays open during runtime that you can output to with the
Debugger. However, the Debugger is not completely static anymore; it's a
singleton now. Just like everything else. Welcome to the club, Debugger.
And yes, this requires the user to call the get function whenever they
use the Debugger. But it's not the end of the world! The new Debugger
has lots of advantages over the old one. For one, it being a singleton
instead of a static class means that it can have its own member
variables and perform fewer calculations during runtime (YAY!). It also
has a cool title on the console window: "RGL Debugger." It looks more
official. Finally, the new Debugger doesn't register as the default
ostream, so std::cout doesn't output to the console, and the user won't
do that by accident. So in my opinion, its actually better than before.
But either way, it has to be like this. So deal with it.