Improved some stuff and FINALLY STARTED ON PHYSICS!
Box2D implementation has begun! Added the PhysicsObject class for using
Box2D physics with an object. Very simple at the moment, but will add
functionality in the future as needed. Also made some more changes to
the Debugger - it now shows the title of the game + " Debugger" instead
of "RGL Debugger." The Debugger also can't be manually initialized /
destroyed by the user anymore: only the Game can do so through the use
of the useDebugging parameter. Also made some changes to the LevelParser
- I moved its member variables to be stored in the parsing Level class.
This enabled me to make the LevelParser class static, once again making
it easier for the user to use.