Added collision callbacks and better scrolling system.
Collision callbacks are now a thing.  PhysicsObjects can implement
onBeginContact and onEndContact functions for defining what happens when
a collision occurs.  These functions provide access to the PhysicsObject
that was collided with (if there was one) and where the collision
occurred.  Also implemented functions in the Level class to provide
automatic scaling/scrolling.  It is now possible to easily get the mouse
coordinates relative to the level itself.