TextBox added and other improvements made.
TextBox class was added for text input. Very minimal at the moment;
improvements will come in the future. Made miscellaneous improvements to
the FontManager - mainly antialiasing and the ability to get the size of
characters and strings.  The InputHandler can handle text input that
takes modifier buttons into account (like Shift). The isTileAt function
in Level now only returns true if the tile is interactive.