New, faster text rendering.
The FontManager class has been overhauled.  Instead of regenerating
textures upon each draw function, textures and information regarding
them can be generated and cached for reuse.  This makes overall
rendering a whole lot faster than it once was.  On-the-fly rendering is
still a thing, though, for situations in which the text being rendered
is changing very rapidly.  However, caching text is almost always
beneficial.  It is easy to do and worth the extra line of code.